rename function for consistency
