+++ aws/lambda/oss-ci-job-queue-time/Makefile
@@ -11,7 +11,7 @@ venv/bin/python:
 
 deployment.zip:
 	mkdir -p deployment
-	cp oss_ci_job_queue_time.py ./deployment/.
+	cp lambda_function.py ./deployment/.
 	pip3.10 install -r requirements.txt -t ./deployment/. --platform manylinux2014_x86_64 --only-binary=:all: --implementation cp --python-version 3.10 --upgrade
 	cd ./deployment && zip -q -r ../deployment.zip .
