+++ aws/lambda/oss-ci-job-queue-time/requirements.txt
@@ -1,2 +1,3 @@
 clickhouse_connect==0.8.5
+boto3==1.35.33
 pytest==7.4.0
